[PT-D] Fix Sharding spec inference to avoid invalid chunk sharding to be inferred as chunkshardingspec
Our previous logic does not fully all corner cases when it comes to imbalance sharding. For example, one is 16 another one is 9. (This cannot be chunk sharding) because the total length is 25 and if it's chunk sharding, it should be 13 and 12. So we need to get the total length first and calculated the expected chunk length to ensure it's indeed a chunk sharding case.

Also added more test cases in unit test.

Differential Revision: [D35417257](https://our.internmc.facebook.com/intern/diff/D35417257/)

ghstack-source-id: 153136085
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75296